Widen the format tag so identifiers cannot collide with its flags.
format_create() packs FORMAT_PANE or FORMAT_WINDOW into the same u_int
as a pane or window identifier, and the tag is only used to key the
format job cache. FORMAT_PANE is 0x80000000 and FORMAT_WINDOW is
0x40000000, so an identifier that reached either bit would be read as
the other kind of tag and share another object's cached job output.

Identifiers are u_int counters that increment once per object created,
so this needs about a billion windows or panes in one server and nobody
has hit it. Move the two flags above 32 bits and make the tag uint64_t
so the identifier range and the flags cannot meet.

+++ format.c
@@ -55,7 +55,7 @@ static void	 format_defaults_winlink(struct format_tree *,
 /* Entry in format job tree. */
 struct format_job {
 	struct client		*client;
-	u_int			 tag;
+	uint64_t		 tag;
 	const char		*cmd;
 	const char		*expanded;
 
@@ -179,7 +179,7 @@ struct format_tree {
 	struct cmdq_item	*item;
 	struct client		*client;
 	int			 flags;
-	u_int			 tag;
+	uint64_t		 tag;
 
 	struct mouse_event	 m;
 
@@ -4276,7 +4276,8 @@ format_create_add_item(struct format_tree *ft, struct cmdq_item *item)
 
 /* Create a new tree. */
 struct format_tree *
-format_create(struct client *c, struct cmdq_item *item, int tag, int flags)
+format_create(struct client *c, struct cmdq_item *item, uint64_t tag,
+    int flags)
 {
 	struct format_tree	*ft;
